split files out and write test for finalise gateway
