+++ ManageATenancyAPI/Gateways/SaveMeeting/SaveEtraMeetingSignOffMeeting/BasePersistenceServiceConfiguration.cs
@@ -1,0 +1,9 @@
+namespace ManageATenancyAPI.Gateways.SaveMeeting.SaveEtraMeetingSignOffMeeting
+{
+    public class BasePersistenceServiceConfiguration
+    {
+        public FileType FileType { get; set; }
+        public string Extension { get; set; }
+        public string ProjectName { get; set; }
+    }
+}
