+++ ManageATenancyAPI/Gateways/SaveMeeting/SaveEtraMeetingSignOffMeeting/FileType.cs
@@ -1,0 +1,10 @@
+namespace ManageATenancyAPI.Gateways.SaveMeeting.SaveEtraMeetingSignOffMeeting
+{
+    public enum FileType
+    {
+        None,
+        Pdf,
+        Email,
+        Jpeg
+    }
+}
