+++ ManageATenancyAPI/Gateways/SaveMeeting/SaveEtraMeetingSignOffMeeting/IJpegPersistenceService.cs
@@ -1,0 +1,7 @@
+namespace ManageATenancyAPI.Gateways.SaveMeeting.SaveEtraMeetingSignOffMeeting
+{
+    public interface IJpegPersistenceService : IPersistenceService
+    {
+
+    }
+}
