+++ ManageATenancyAPI/Gateways/SaveMeeting/SaveEtraMeetingSignOffMeeting/IPersistenceService.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+
+namespace ManageATenancyAPI.Gateways.SaveMeeting.SaveEtraMeetingSignOffMeeting
+{
+    public interface IPersistenceService
+    {
+        Task SaveAsync(string id, byte[] payload);
+        Task<byte[]> GetAsync(string id);
+    }
+}
