+++ ManageATenancyAPI/Gateways/SaveMeeting/SaveEtraMeetingSignOffMeeting/JpegPersistenceServiceConfiguration.cs
@@ -1,0 +1,7 @@
+namespace ManageATenancyAPI.Gateways.SaveMeeting.SaveEtraMeetingSignOffMeeting
+{
+    public class JpegPersistenceServiceConfiguration : BasePersistenceServiceConfiguration
+    {
+
+    }
+}
